+++ test/test.namespaces.js
@@ -34,8 +34,8 @@ describe("Twig.js Namespaces ->", function() {
     });
   });
 
-    it("should support namespaces defined with @", function(done) {
-    	twig({
+  it("should support namespaces defined with @", function(done) {
+   twig({
 			namespaces: { 'test': 'test/templates/namespaces/' },
 			path: 'test/templates/namespaces_@.twig',
 			load: function(template) {
@@ -46,9 +46,9 @@ describe("Twig.js Namespaces ->", function() {
 				}).should.equal("namespaces");
 
 				done();
-            }
-    	});
+      }
     });
+	});
 
   it("should support namespaces defined with @ in source method", function(done) {
     twig({
@@ -66,5 +66,19 @@ describe("Twig.js Namespaces ->", function() {
     });
   });
 
+	it("should support non-namespaced includes with namespaces configured", function(done) {
+    	twig({
+			namespaces: { 'test': 'test/templates/namespaces/' },
+			path: 'test/templates/namespaces_without_namespace.twig',
+			load: function(template) {
+				// Render the template
+				template.render({
+				    test: "yes",
+				    flag: true
+				}).should.equal("namespaces\nnamespaces");
 
+				done();
+            }
+    	});
+    });
 });